Move putNonEmptySfu to SetupFutureUsageUtils.kt
