+++ payments-core/src/main/java/com/stripe/android/utils/SetupFutureUsageUtils.kt
@@ -11,3 +11,14 @@ fun ConfirmPaymentIntentParams.SetupFutureUsage?.hasIntentToSetup() = when (this
     ConfirmPaymentIntentParams.SetupFutureUsage.None -> false
     null -> false
 }
+
+private const val PARAM_SETUP_FUTURE_USAGE = "setup_future_usage"
+internal fun MutableMap<String, Any>.putNonEmptySfu(sfu: ConfirmPaymentIntentParams.SetupFutureUsage?) {
+    sfu.takeIf {
+        // Empty values are an attempt to unset a parameter;
+        // however, setup_future_usage cannot be unset.
+        it != null && it != ConfirmPaymentIntentParams.SetupFutureUsage.Blank
+    }?.let {
+        put(PARAM_SETUP_FUTURE_USAGE, it.code)
+    }
+}
